+++ pkg/gen/input/dependabot/internal/params/key_common.go
@@ -7,11 +7,3 @@ import (
 func Header(comment string) string {
 	return internal.Header(comment)
 }
-
-func Package(p Params) string {
-	return internal.Package(p.Dir)
-}
-
-func RegenerableFileName(p Params, suffix string) string {
-	return internal.RegenerableFileName(p.Dir, suffix)
-}