+++ pkg/gen/input/renovate/internal/params/key.go
@@ -1,0 +1,13 @@
+package params
+
+func Interval(p Params) string {
+	return p.Interval
+}
+
+func Language(p Params) string {
+	return p.Language
+}
+
+func Reviewer(p Params) string {
+	return p.Reviewer
+}