+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Add `devctl gen renovate` command.
+
 - Dependencies: use github.com/gorilla/websocket version v1.4.2
 
 ## [4.7.0] - 2021-07-08